+++ src/DotNext/Text/Base64Decoder.Unicode.cs
@@ -42,7 +42,7 @@ namespace DotNext.Text
         private void CopyAndDecode(ReadOnlySpan<char> chars, IBufferWriter<byte> output)
         {
             var newSize = reservedBufferSize + chars.Length;
-            using var tempBuffer = newSize <= MemoryRental<char>.StackallocThreshold ? stackalloc char[newSize] : new MemoryRental<char>(newSize);
+            using var tempBuffer = (uint)newSize <= MemoryRental<char>.StackallocThreshold ? stackalloc char[newSize] : new MemoryRental<char>(newSize);
             ReservedChars.Slice(0, reservedBufferSize).CopyTo(tempBuffer.Span);
             chars.CopyTo(tempBuffer.Span.Slice(reservedBufferSize));
             DecodeCore(tempBuffer.Span, output);
@@ -129,7 +129,7 @@ namespace DotNext.Text
             where TConsumer : notnull, IReadOnlySpanConsumer<byte>
         {
             var newSize = reservedBufferSize + chars.Length;
-            using var tempBuffer = newSize <= MemoryRental<char>.StackallocThreshold ? stackalloc char[newSize] : new MemoryRental<char>(newSize);
+            using var tempBuffer = (uint)newSize <= MemoryRental<char>.StackallocThreshold ? stackalloc char[newSize] : new MemoryRental<char>(newSize);
             ReservedChars.Slice(0, reservedBufferSize).CopyTo(tempBuffer.Span);
             chars.CopyTo(tempBuffer.Span.Slice(reservedBufferSize));
             DecodeCore(tempBuffer.Span, output);
